[pods] Simplify: shared manifest folder-name validation and test fixtures
